filters work

+++ guidebook/guidebook_app/utils.py
@@ -22,7 +22,7 @@ def stringify_list(col, df) -> pd.DataFrame:
 
 
 def repopulate_list(col, df) -> None:
-    df[col] = df[col].apply(lambda x: [float(y) for y in x.split(",")]).astype(object)
+    df[col] = df[col].apply(lambda x: [int(y) for y in x.split(",")]).astype(object)
 
 
 def stash_dataframe(df, int64_cols=[], list_cols=[]) -> list:
